Make the amount control buttons functional
Add a set amount to each cartItem and add it to
createItemRightSections parameters.
Change input type of item amount to number and add readOnly (for now).
Create itemPrice with "let" so users can control item amount.
Add click eventlisteners on increase and reduce buttons and call a function.
Create said function which will be used on the buttons.
The function will, onclick, fetch info of the nearest cartItem to the click
and manipulate the DOM and the object it is related to.