Rewrite validateStringForNumbersOrSymbols and validateOrderForm
Change validateStringForNumbersOrSymbols name to validateInputField.
Rewrite the code within to take 4 arguments. The negate arg will
check if the if statement should use the NOT operator.
The regex arg is pass whatever regex desired. The input arg
is the input element. The msg arg is the message you want to
give to the user when validation fails.
validateOrderForm now defines regex in variables for readability
and calls the validateInputField on retrieved input fields.