Fix the increaseOrDecrease logic
Create an empty variable that will hold
the default price.
Select all products on DOM and go through
each product and create a variable that is the
products name and check if it is the same as the
cartItems name. Then add the products price into
the defaultProductPrice.
Now the obj.price will adapt by multiplying the
obj.amount with the defaultProductPrice.