Make the reset button functional
Add a click eventlistener to the reset button which will
reset the cart in DOM, remove everything from cartArray
and update the total price in the form.
Make changes in the updateTotalPrice function that
will check if the cartArray is empty and add an empty default text
if it's checked out. Also switch the position of the deliveryFee
so it gets reset with each execution of the updateTotalPrice function.