Organizing and logic changes
Move global variables to the top.
In the document addeventlistener, check if the
submitOrderBtn is disabled and style the background to grey.
In the formIsValid function, select all .form-row since the previous
selector only selected those with inline display flex style.
Create an array that will contain the form rows with display flex.
Go through each row in formRows, and check if they have display flex
and push into the array.
Go through the each row flexFormRows and select all required inputs.
Go through each of the required inputs and check if they are empty and
add to the "empty" variable.
If the "if statement" is met, enable submit btn and make its background
green. Else disable the btn and make the background of it, grey.